// packages/storefront/src/app/common/generated-types.ts
export type Maybe<T> = T;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Channel = Node & {
  __typename?: 'Channel';
  code: Scalars['String'];
  createdAt: Scalars['DateTime'];
  currencyCode: CurrencyCode;
  customFields?: Maybe<Scalars['JSON']>;
  defaultLanguageCode: LanguageCode;
  defaultShippingZone?: Maybe<Zone>;
  defaultTaxZone?: Maybe<Zone>;
  id: Scalars['ID'];
  pricesIncludeTax: Scalars['Boolean'];
  token: Scalars['String'];
  updatedAt: Scalars['DateTime'];
};
// ...
export type IneligiblePaymentMethodError = ErrorResult & {
  __typename?: 'IneligiblePaymentMethodError';
  eligibilityCheckerMessage?: Maybe<Scalars['String']>;
  errorCode: ErrorCode;
  message: Scalars['String'];
};
// ...
export type NoActiveOrderError = ErrorResult & {
  __typename?: 'NoActiveOrderError';
  errorCode: ErrorCode;
  message: Scalars['String'];
};

/**
 * Returned if `authOptions.requireVerification` is set to `true` (which is the default)
 * and an unverified user attempts to authenticate.
 */
export type NotVerifiedError = ErrorResult & {
  __typename?: 'NotVerifiedError';
  errorCode: ErrorCode;
  message: Scalars['String'];
};

export type Order = Node & {
  __typename?: 'Order';
  /** An order is active as long as the payment process has not been completed */
  active: Scalars['Boolean'];
  /**
   * Order-level adjustments to the order total, such as discounts from promotions
   * @deprecated Use `discounts` instead
   */
  adjustments: Array<Adjustment>;
  billingAddress?: Maybe<OrderAddress>;
  /** A unique code for the Order */
  code: Scalars['String'];
  /** An array of all coupon codes applied to the Order */
  couponCodes: Array<Scalars['String']>;
  createdAt: Scalars['DateTime'];
  currencyCode: CurrencyCode;
  customFields?: Maybe<Scalars['JSON']>;
  customer?: Maybe<Customer>;
  discounts: Array<Discount>;
  fulfillments?: Maybe<Array<Fulfillment>>;
  history: HistoryEntryList;
  id: Scalars['ID'];
  lines: Array<OrderLine>;
  /**
   * The date & time that the Order was placed, i.e. the Customer
   * completed the checkout and the Order is no longer "active"
   */
  orderPlacedAt?: Maybe<Scalars['DateTime']>;
  payments?: Maybe<Array<Payment>>;
  /** Promotions applied to the order. Only gets populated after the payment process has completed. */
  promotions: Array<Promotion>;
  shipping: Scalars['Int'];
  shippingAddress?: Maybe<OrderAddress>;
  shippingLines: Array<ShippingLine>;
  shippingWithTax: Scalars['Int'];
  state: Scalars['String'];
  /**
   * The subTotal is the total of all OrderLines in the Order. This figure also includes any Order-level
   * discounts which have been prorated (proportionally distributed) amongst the OrderItems.
   * To get a total of all OrderLines which does not account for prorated discounts, use the
   * sum of `OrderLine.discountedLinePrice` values.
   */
  subTotal: Scalars['Int'];
  /** Same as subTotal, but inclusive of tax */
  subTotalWithTax: Scalars['Int'];
  /**
   * Surcharges are arbitrary modifications to the Order total which are neither
   * ProductVariants nor discounts resulting from applied Promotions. For example,
   * one-off discounts based on customer interaction, or surcharges based on payment
   * methods.
   */
  surcharges: Array<Surcharge>;
  /** A summary of the taxes being applied to this Order */
  taxSummary: Array<OrderTaxSummary>;
  /** Equal to subTotal plus shipping */
  total: Scalars['Int'];
  totalQuantity: Scalars['Int'];
  /** The final payable amount. Equal to subTotalWithTax plus shippingWithTax */
  totalWithTax: Scalars['Int'];
  updatedAt: Scalars['DateTime'];
};
// ...
export type OrderPaymentStateError = ErrorResult & {
  __typename?: 'OrderPaymentStateError';
  errorCode: ErrorCode;
  message: Scalars['String'];
};

/** Returned if there is an error in transitioning the Order state */
export type OrderStateTransitionError = ErrorResult & {
  __typename?: 'OrderStateTransitionError';
  errorCode: ErrorCode;
  fromState: Scalars['String'];
  message: Scalars['String'];
  toState: Scalars['String'];
  transitionError: Scalars['String'];
};
// ...
export type PaymentDeclinedError = ErrorResult & {
  __typename?: 'PaymentDeclinedError';
  errorCode: ErrorCode;
  message: Scalars['String'];
  paymentErrorMessage: Scalars['String'];
};

/** Returned when a Payment fails due to an error. */
export type PaymentFailedError = ErrorResult & {
  __typename?: 'PaymentFailedError';
  errorCode: ErrorCode;
  message: Scalars['String'];
  paymentErrorMessage: Scalars['String'];
};
// ...
export type Query = {
  __typename?: 'Query';
  /** The active Channel */
  activeChannel: Channel;
  /** The active Customer */
  activeCustomer?: Maybe<Customer>;
  /**
   * The active Order. Will be `null` until an Order is created via `addItemToOrder`. Once an Order reaches the
   * state of `PaymentApproved` or `PaymentSettled`, then that Order is no longer considered "active" and this
   * query will once again return `null`.
   */
  activeOrder?: Maybe<Order>;
  /** An array of supported Countries */
  availableCountries: Array<Country>;
  /** Returns a Collection either by its id or slug. If neither 'id' nor 'slug' is speicified, an error will result. */
  collection?: Maybe<Collection>;
  /** A list of Collections available to the shop */
  collections: CollectionList;
  /** Returns a list of payment methods and their eligibility based on the current active Order */
  eligiblePaymentMethods: Array<PaymentMethodQuote>;
  /** Returns a list of eligible shipping methods based on the current active Order */
  eligibleShippingMethods: Array<ShippingMethodQuote>;
  /** Returns information about the current authenticated User */
  me?: Maybe<CurrentUser>;
  /** Returns the possible next states that the activeOrder can transition to */
  nextOrderStates: Array<Scalars['String']>;
  /**
   * Returns an Order based on the id. Note that in the Shop API, only orders belonging to the
   * currently-authenticated User may be queried.
   */
  order?: Maybe<Order>;
  /**
   * Returns an Order based on the order `code`. For guest Orders (i.e. Orders placed by non-authenticated Customers)
   * this query will only return the Order within 2 hours of the Order being placed. This allows an Order confirmation
   * screen to be shown immediately after completion of a guest checkout, yet prevents security risks of allowing
   * general anonymous access to Order data.
   */
  orderByCode?: Maybe<Order>;
  /** Get a Product either by id or slug. If neither 'id' nor 'slug' is speicified, an error will result. */
  product?: Maybe<Product>;
  /** Get a list of Products */
  products: ProductList;
  /** Search Products based on the criteria set by the `SearchInput` */
  search: SearchResponse;
};
// ...
export type PaymentInput = {
  /**
   * This field should contain arbitrary data passed to the specified PaymentMethodHandler's `createPayment()` method
   * as the "metadata" argument. For example, it could contain an ID for the payment and other
   * data generated by the payment provider.
   */
  metadata: Scalars['JSON'];
  /** This field should correspond to the `code` property of a PaymentMethodHandler. */
  method: Scalars['String'];
};
// ...
  export type Variables = UpdateAddressMutationVariables;
// ...
  export type ShippingAddress = NonNullable<NonNullable<GetOrderQuery['orderByCode']>['shippingAddress']>;
// ...
export namespace AddPayment {
  export type Variables = AddPaymentMutationVariables;
  export type Mutation = AddPaymentMutation;
  export type AddPaymentToOrder = NonNullable<AddPaymentMutation['addPaymentToOrder']>;
}
// ...
  export type TransitionOrderToState = NonNullable<TransitionToAddingItemsMutation['transitionOrderToState']>;
// ...
  export type ActiveOrder = NonNullable<GetShippingAddressQuery['activeOrder']>;
// ...
export namespace TransitionToArrangingPayment {
  export type Variables = TransitionToArrangingPaymentMutationVariables;
  export type Mutation = TransitionToArrangingPaymentMutation;
  export type TransitionOrderToState = NonNullable<
    TransitionToArrangingPaymentMutation['transitionOrderToState']
  >;
}

export namespace GetOrderForCheckout {
  export type Variables = GetOrderForCheckoutQueryVariables;
  export type Query = GetOrderForCheckoutQuery;
  export type ActiveOrder = NonNullable<GetOrderForCheckoutQuery['activeOrder']>;
  export type ShippingAddress = NonNullable<
    NonNullable<GetOrderForCheckoutQuery['activeOrder']>['shippingAddress']
  >;
}
// ...
export namespace Country {
  export type Fragment = CountryFragment;
}
// ...
export type UpdateAddressMutationVariables = Exact<{
  input: UpdateAddressInput;
}>;
// ...
export type AddPaymentMutationVariables = Exact<{
  input: PaymentInput;
}>;

export type AddPaymentMutation = { __typename?: 'Mutation' } & {
  addPaymentToOrder:
    | ({ __typename?: 'IneligiblePaymentMethodError' } & ErrorResult_IneligiblePaymentMethodError_Fragment)
    | ({ __typename?: 'NoActiveOrderError' } & ErrorResult_NoActiveOrderError_Fragment)
    | ({ __typename?: 'Order' } & CartFragment)
    | ({ __typename?: 'OrderPaymentStateError' } & ErrorResult_OrderPaymentStateError_Fragment)
    | ({ __typename?: 'OrderStateTransitionError' } & ErrorResult_OrderStateTransitionError_Fragment)
    | ({ __typename?: 'PaymentDeclinedError' } & ErrorResult_PaymentDeclinedError_Fragment)
    | ({ __typename?: 'PaymentFailedError' } & ErrorResult_PaymentFailedError_Fragment);
};
// ...
export type TransitionToArrangingPaymentMutationVariables = Exact<{ [key: string]: never }>;

export type TransitionToArrangingPaymentMutation = { __typename?: 'Mutation' } & {
  transitionOrderToState?: Maybe<
    | ({ __typename?: 'Order' } & CartFragment)
    | ({ __typename?: 'OrderStateTransitionError' } & ErrorResult_OrderStateTransitionError_Fragment)
  >;
};

export type GetOrderForCheckoutQueryVariables = Exact<{ [key: string]: never }>;

export type GetOrderForCheckoutQuery = { __typename?: 'Query' } & {
  activeOrder?: Maybe<
    { __typename?: 'Order' } & {
      shippingAddress?: Maybe<{ __typename?: 'OrderAddress' } & OrderAddressFragment>;
    } & CartFragment
  >;
};
// ...
export type CartFragment = { __typename?: 'Order' } & Pick<
  Order,
  | 'id'
  | 'code'
  | 'state'
  | 'active'
  | 'totalQuantity'
  | 'subTotal'
  | 'subTotalWithTax'
  | 'total'
  | 'totalWithTax'
  | 'shipping'
  | 'shippingWithTax'
> & {
    lines: Array<
      { __typename?: 'OrderLine' } & Pick<
        OrderLine,
        | 'id'
        | 'unitPrice'
        | 'unitPriceWithTax'
        | 'quantity'
        | 'linePriceWithTax'
        | 'discountedLinePriceWithTax'
      > & {
          featuredAsset?: Maybe<{ __typename?: 'Asset' } & AssetFragment>;
          productVariant: { __typename?: 'ProductVariant' } & Pick<ProductVariant, 'id' | 'name'>;
          discounts: Array<
            { __typename?: 'Discount' } & Pick<
              Discount,
              'amount' | 'amountWithTax' | 'description' | 'adjustmentSource' | 'type'
            >
          >;
        }
    >;
    shippingLines: Array<
      { __typename?: 'ShippingLine' } & Pick<ShippingLine, 'priceWithTax'> & {
          shippingMethod: { __typename?: 'ShippingMethod' } & Pick<
            ShippingMethod,
            'id' | 'code' | 'name' | 'description'
          >;
        }
    >;
    discounts: Array<
      { __typename?: 'Discount' } & Pick<
        Discount,
        'amount' | 'amountWithTax' | 'description' | 'adjustmentSource' | 'type'
      >
    >;
  };
// ...
type ErrorResult_IneligiblePaymentMethodError_Fragment = {
  __typename?: 'IneligiblePaymentMethodError';
} & Pick<IneligiblePaymentMethodError, 'errorCode' | 'message'>;
// ...
type ErrorResult_NoActiveOrderError_Fragment = { __typename?: 'NoActiveOrderError' } & Pick<
  NoActiveOrderError,
  'errorCode' | 'message'
>;
// ...
type ErrorResult_OrderPaymentStateError_Fragment = { __typename?: 'OrderPaymentStateError' } & Pick<
  OrderPaymentStateError,
  'errorCode' | 'message'
>;

type ErrorResult_OrderStateTransitionError_Fragment = { __typename?: 'OrderStateTransitionError' } & Pick<
  OrderStateTransitionError,
  'errorCode' | 'message'
>;
// ...
type ErrorResult_PaymentDeclinedError_Fragment = { __typename?: 'PaymentDeclinedError' } & Pick<
  PaymentDeclinedError,
  'errorCode' | 'message'
>;

type ErrorResult_PaymentFailedError_Fragment = { __typename?: 'PaymentFailedError' } & Pick<
  PaymentFailedError,
  'errorCode' | 'message'
>;